feat: solve detail page

+++ app/src/main/java/com/kom/cryptospot/data/source/network/model/coindetail/CurrentPrice.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 @Keep
 data class CurrentPrice(
     @SerializedName("idr")
-    val idr: Int?,
+    val idr: Double?,
     @SerializedName("usd")
-    val usd: Int?,
+    val usd: Double?,
 )
